feat(scripts): polimento da logica de dataset
